Improve the flask example, fixes #10
